refactor: 유틸 및 커스텀 훅 분리
